requirements.txt
fastapi
uvicorn[standard]
transformers==4.41.2
torch
scipy
numpy
yt-dlp
openai-whisper
librosa
soundfile
python-multipart
pydantic-settings 